Separate testbed lifecycle ownership and capability admission
